rename atlas_utils to atlas
