+++ databuilder/databuilder/models/atlas_entity.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from collections import namedtuple
-from common.amundsen_common.utils.atlas_utils import AtlasSerializedEntityFields
+from common.amundsen_common.utils.atlas import AtlasSerializedEntityFields
 
 AtlasEntity = namedtuple(
     'AtlasEntity',
