+++ databuilder/databuilder/models/dashboard/dashboard_table.py
@@ -10,7 +10,7 @@ from typing import (
 from amundsen_rds.models import RDSModel
 from amundsen_rds.models.dashboard import DashboardTable as RDSDashboardTable
 
-from amundsen_common.utils.atlas_utils import AtlasDashboardTypes, AtlasTableKey, AtlasRelationshipTypes
+from amundsen_common.utils.atlas import AtlasDashboardTypes, AtlasTableKey, AtlasRelationshipTypes
 from databuilder.models.atlas_entity import AtlasEntity
 from databuilder.models.atlas_relationship import AtlasRelationship
 from databuilder.models.atlas_serializable import AtlasSerializable
