+++ metadata/metadata_service/proxy/atlas_proxy.py
@@ -20,12 +20,12 @@ from amundsen_common.models.table import (Badge, Column,
                                           ResourceReport, Stat, Table, Tag,
                                           User, Watermark)
 from amundsen_common.models.user import User as UserEntity
-from amundsen_common.utils.atlas_utils import (AtlasColumnKey,
-                                               AtlasCommonParams,
-                                               AtlasCommonTypes,
-                                               AtlasDashboardTypes,
-                                               AtlasStatus, AtlasTableKey,
-                                               AtlasTableTypes)
+from amundsen_common.utils.atlas import (AtlasColumnKey,
+                                         AtlasCommonParams,
+                                         AtlasCommonTypes,
+                                         AtlasDashboardTypes,
+                                         AtlasStatus, AtlasTableKey,
+                                         AtlasTableTypes)
 from apache_atlas.client.base_client import AtlasClient
 from apache_atlas.model.glossary import (AtlasGlossary, AtlasGlossaryHeader,
                                          AtlasGlossaryTerm)
